requirements.txt
pandas~=2.2.3
apache-airflow~=2.10.3
confluent-kafka~=2.6.1
pyspark~=4.0.0.dev2
azure-core~=1.32.0
kafka~=1.3.5
kafka-python~=2.0.2